+++ lisp/eab-org.el
@@ -77,7 +77,7 @@
 (setq org-archive-default-command 'org-archive-set-tag)
 
 (setq org-tags-exclude-from-inheritance (quote ("crypt")))
-(org-crypt-use-before-save-magic)
+;; (org-crypt-use-before-save-magic)
 (setq org-crypt-key nil)
 
 (setq